requirements.txt
eventlet>=0.33.3
flask>=2.3.2
flask-socketio>=5.3.4
flask-cors>=3.0.10
google-cloud-speech>=2.19.0
gunicorn
 
